kafka/requirements.txt
tweepy
kafka-python
psycopg2-binary

